Use `@CompileStatic` where possible to increase performance.
Slight speed increase, e.g. down from 1300 ms to 1100 ms (15% faster).
The HTML output will probably get much less of this benefit, because that one relies a lot of dynamic method calls.

Upgraded to latest Groovy to benefit from `@CompileStatic` fixes.

In overall it was just an experiment, but it ended up useful enough, so I commmitted it. The goal was to have
the performance of Java but without its verbosity where possible.