Require Java 11 for compilation and as a runtime environment.
According to Groovy 3's release notes:

> Groovy 3.0 requires JDK9+ to build and JDK8 is the minimum version of the JRE that we support.

My app somehow still works, and I did not notice issues when I made the switch to Groovy 3, and probably because I do not use
functionality that is specific to JDK 9. But since JDK 8 and 11 are both LTS versions, I chose to require Java 11. It should not
be a big problem as of today, in my opinion, so for simplicity I chose not to compile for Java 8 runtime.

I also upgraded Gradle to version 7, which also switches to Groovy 3, to keep things more aligned.