Removed support for the `-w` argument option.
The problem with this kind of options is that they are specific to certain types of files, and it is only a matter of time until
we see many of such options.

Maybe in the future I will introduce some plugin capabilities. For now I will focus on core functionalities.