Allow relative paths for included config files.
This change makes it easier to reuse common configuration files, although with an unintended limitation I will open a issue for:
once a property like `maxContentLines` is set, you cannot override it later.

Also wrote a test to check that included config files are indeed loaded and their settings are taken into account.